feat(row-bridge): assemble complex collection columns on the primary SELECT read path
The primary SELECT path (router.rs -> bridge::partition_to_rows_with_storage_mapping,
~15 call sites) funnels through row::decode_output_row, which lived in the
lower ferrosa-row-bridge crate and decoded each cell as if it were the whole
column -- so a complex (per-element) collection read back as a single element
or empty.

Relocate the collection-cell logic (build_collection_cells / assemble_collection
/ list_cell_path / timeuuid_time + types) down from ferrosa-cql::collection_cells
into ferrosa-row-bridge::collection, where its encode_value/decode_value codecs
and ferrosa_common::reconcile already live; ferrosa-cql::collection_cells is now
a re-export shim so existing call sites resolve unchanged.

Add one shared assemble_column_cells(col_type, cells, now_secs) helper that both
read paths use: complex columns reconcile per path (CRDT LWW so a removed element
cannot resurrect), drop non-live cells (tombstoned / TTL-expired), and assemble;
a single path=None cell (scalar OR legacy whole-value collection) decodes as the
whole value (backward-compatible lazy dual-read). Route BOTH decode_output_row
(primary path) and partition_to_rows_with_metadata (metadata variant) through it,
eliminating the assembly duplication the crate's FMEA note warns against. Assembly
failures log loudly rather than silently returning None.

3 new row-bridge tests (list append-order, set-removal reconcile, legacy
whole-value); ferrosa-cql 1026 lib tests + 17 row-bridge tests green; clippy clean.

crdt-collections increment 3, part D-read-2. Refs t_83c4f093.
